Character records backend improvements and minor UI fixes (#159)
* Fix exception for bad record markup

* Provide a default title for unnamed record entries

* Switch to using integer keys for character records

This should be *far* most robust and fix quite a few of the issues we
have been having with records. A component is added to the entity to be
able to lookup the key in things like the admin erase method.

* Use white color for everything in records console

* Remove character records from station on cryosleep

(cherry picked from commit a04a166180633801af5c656b94c9c74d147f3a58)