Port over ResetParacusia reagent effect from Cosmatic Drift (#42)
* Update medicine.yml

* Anti-Psych RP Chem (#105)

* Add Saprodipine RP Chem

Saprodipine is intended to be an anti psych medicine. Currently it does
not suppress any of the hallucination traits. The name as far as I know
is completely random. This was done to ensure it does not overlap with
something from real life.

* Saprodipine suppresses paracusia

Add the ResetParacusia reagent effect which is basically ResetNarcolepsy
but for Paracusia, Saprodipine now applies this effect. Modifed
ParacusiaComponent and ParacusiaSystem to support delaying Paracusia
incidents.

* Typofixes and add messages to Saprodipine OD

* Remove merge conflict traces

* Rename Kate to Ghazet everywhere.

* Listinv to gridinv

* Antidepressants and Anxiety RP Meds (#76)

* add everything

* fix test fail

---------

Co-authored-by: dffdff2423 <57052305+dffdff2423@users.noreply.github.com>
Co-authored-by: CrigCrag <137215465+CrigCrag@users.noreply.github.com>