Fix accidental dropping of CD columns
The `DropColumn` migration with SQLite requires the entire table to be
rebuilt. Entity Framework rebuilds the table according to what's in the
.Designer.cs file. Since that was generated in a repo that doesn't have
the CD columns, Entity Framework drops those columns too.

The solution is to add the CD columns to any migration that causes the
`profile` table to be rebuilt. This is remarkably ugly, but effective.