Add two new alert levels - orange and white.
Added two new alert levels with their custom alert sounds to the game and communications console that players can select: orange and white.

The orange alert serves as a middle-ground between blue alert and red alert, focusing on hostile or unknown life forms on board the station. With it's announcement, it states: "There are currently multiple unknown or hostile life forms on board the station. All crewmembers are advised to stay in their departments and to report any aggressive entities to security." while the PDA instructions state: "Crewmembers are advised to be vigilant and report hostile life forms to security."

While the white alert concentrates on supernatural or otherwise unexplained phenomena which has become more common in the Umbra Sector stations. The announcement states: "There is currently a supernatural or eldritch level abnormality on the station. All crewmembers are to remain vigilant, seek shelter, listen to command orders and to report any supernatural phenomena to security." whilst the PDA instructions state: "Crewmembers are advised to be remain in their departments and to report any supernatural sightings to security."

Both of the alerts alter the SO A-2 subsection of the Standard Operating Procedure. For the orange alert, it is:
- Code Orange: Commanding Officer > Head of Security > Executive Officer > Security department > Relevant head-of-department > Relevant department senior > Regular crew
And for the white it's:
- Code White: Commanding Officer > Research Director > Executive Officer > Head Of Security > Chaplain >  Security department > Relevant head-of-department > Relevant department senior > Regular crew

It may change in the future depending on how it plays out but this is what I currently have in mind.